+++ core/src/main/java/hu/iit/uni/miskolc/swtest/model/Library.java
@@ -1,8 +1,8 @@
 package hu.iit.uni.miskolc.swtest.model;
 
-import hu.iit.uni.miskolc.swtest.exceptions.BookListNullException;
-import hu.iit.uni.miskolc.swtest.exceptions.IdNotValidException;
-import hu.iit.uni.miskolc.swtest.exceptions.NameIsEmptyException;
+import hu.iit.uni.miskolc.swtest.model.exceptions.BookListNullException;
+import hu.iit.uni.miskolc.swtest.model.exceptions.IdNotValidException;
+import hu.iit.uni.miskolc.swtest.model.exceptions.NameIsEmptyException;
 
 import java.util.Collection;
