+++ core/src/main/java/hu/iit/uni/miskolc/swtest/model/exceptions/AvailableNotValidException.java
@@ -1,4 +1,4 @@
-package hu.iit.uni.miskolc.swtest.exceptions;
+package hu.iit.uni.miskolc.swtest.model.exceptions;
 
 public class AvailableNotValidException extends Exception {
     public AvailableNotValidException() {
