+++ service/src/main/java/hu/iit/uni/miskolc/swtest/service/LibrarianManagerImpl.java
@@ -5,10 +5,10 @@ import hu.iit.uni.miskolc.swtest.dao.ReaderDao;
 import hu.iit.uni.miskolc.swtest.dao.RequestDao;
 import hu.iit.uni.miskolc.swtest.dao.exceptions.BookEntryAlreadyAddedException;
 import hu.iit.uni.miskolc.swtest.dao.exceptions.BookEntryNotFoundException;
-import hu.iit.uni.miskolc.swtest.exceptions.IdNotValidException;
 import hu.iit.uni.miskolc.swtest.model.Book;
 import hu.iit.uni.miskolc.swtest.model.Reader;
 import hu.iit.uni.miskolc.swtest.model.Request;
+import hu.iit.uni.miskolc.swtest.model.exceptions.IdNotValidException;
 import hu.iit.uni.miskolc.swtest.service.exceptions.BookAlreadyAddedException;
 import hu.iit.uni.miskolc.swtest.service.exceptions.BookDoesNotExistException;
